requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
openai==1.3.0
aiohttp==3.9.1
asyncio
lxml==4.9.3 
